Message generation and offline database proof of concept
Added a small repository of messages and a method of retrieving them
from alarm manager. The a button is added to testDB Fragment for testing
this.

The alarm manager is eventually responsible for initiating the countdown
for the reminder message. The timing and content of the message is
determined by functions in ReminderOracle, using the statistics of the
user.

ReminderOracle currently just dictates type of message and anger level,
which is used as parameters to retrieve a message from the offline
database.

Also made some changes to the newsfeed and daypicker fragments for UI
purposes.