Preliminary styling of notifications
Updated the boring notification to a colorful new one that expands. WIll
think more about the styles.

Found a bug where we write day visit before updating to a new day. So if
you don't open the app for a few days and visit the gym, it will update
an old day record. My fix is to use the onStart update code to update
date before writing.

Removed redundant initialization of geofence/alarmmanager in onCreate,
since they are already done in bootReceiver