Fixed bug in alarm manager not registering on intial setup
A while ago I removed the AlarmManager set in onCreate, so the only
alarm was set in BootReceiver. Unfortunately this meant that the first
install didn't set the alarm manager correctly - it would only work
after a reboot.

This change reintroduces the alarm manager set in oncreate, but with a
checking mechanism to prevent multiple sets (although I dont know if
that is an issue).