Major bug fixes in Gym Vist registration
Users were reporting instances when turning on phone during gym visit
would cause a new visit to be reported. While some symptoms of this bug
were ameliorated by the continuous notification change, it still ended
up causing the first visit to be truncated.

The problem as it turns out was that I was reregistering Geofence in
Oncreate of the main activity. This was not intentional.

After some testing, the app performs much better when GeofenceController
initiation is performed only by boot and firsttime.